trying to get view to work
I tried a lot of things, so that's reflected in the sheer number of changes. Moreover, some of this is from my attempts to get the mySQL database to work and such. The database, model, and controller should (emphasis on should) be working now, but I can't seem to get the view to work.